// api/src/modules/users/dto/create-user.dto.ts
import { z } from "zod";

export const createUserDtoSchema = z.object({
  name: z.string().min(2, "Name must be at least 2 characters"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(8, "Password must be at least 8 characters"),
});

export type CreateUserDto = z.infer<typeof createUserDtoSchema>;